Add pattern 07 to perfect set
This "1-second blue time" pattern gets you two of four 1600s on the six 1-second blue time boards and sets you up nicely for the third. The pattern ends when you reach the top left corner of the board. At this point you're set up for the third 1600, and with decent timing you'll get it.